Moved Param groups to seperate components
